+++ src/ai_marketplace_monitor/config.py
@@ -1,13 +1,20 @@
 import tomllib
 
+from .utils import merge_dicts
+
 
 class Config:
-    def __init__(self, config_file):
-        try:
-            with open(config_file, "rb") as f:
-                self.config = tomllib.load(f)
-        except tomllib.TOMLDecodeError as e:
-            raise ValueError(f"Error parsing config file {config_file}: {e}") from e
+    def __init__(self, config_files):
+        configs = []
+        for config_file in config_files:
+            try:
+                with open(config_file, "rb") as f:
+                    configs.append(tomllib.load(f))
+            except tomllib.TOMLDecodeError as e:
+                raise ValueError(f"Error parsing config file {config_file}: {e}") from e
+        #
+        # merge the list of configs into a single dictionary, including dictionaries in the values
+        self.config = merge_dicts(configs)
         self.validate()
 
     def validate(self) -> None:
@@ -29,7 +36,7 @@ class Config:
         for marketplace_name, marketplace_config in self.config["marketplace"].items():
             if marketplace_name not in supported_marketplaces:
                 raise ValueError(
-                    f"Marketplace {marketplace_name} is not supported. Supported marketplaces are: {supported_marketplaces.keys()}"
+                    f"Marketplace [blue]{marketplace_name}[/blue] is not supported. Supported marketplaces are: {supported_marketplaces.keys()}"
                 )
             marketplace_class = supported_marketplaces[marketplace_name]
             marketplace_class.validate(marketplace_config)
@@ -41,11 +48,13 @@ class Config:
             if "marketplace" in item_config:
                 if item_config["marketplace"] not in self.config["marketplace"]:
                     raise ValueError(
-                        f"Item {item_name} specifies a marketplace that does not exist."
+                        f"Item [blue]{item_name}[/blue] specifies a marketplace that does not exist."
                     )
 
             if "keywords" not in item_config:
-                raise ValueError(f"Item {item_name} does not contain a keywords to search.")
+                raise ValueError(
+                    f"Item [blue]{item_name}[/blue] does not contain a keywords to search."
+                )
             #
             if isinstance(item_config["keywords"], str):
                 item_config["keywords"] = [item_config["keywords"]]
@@ -53,13 +62,21 @@ class Config:
             if not isinstance(item_config["keywords"], list) or not all(
                 isinstance(x, str) for x in item_config["keywords"]
             ):
-                raise ValueError(f"Item {item_name} keywords must be a list.")
+                raise ValueError(f"Item [blue]{item_name}[/blue] keywords must be a list.")
             if len(item_config["keywords"]) == 0:
-                raise ValueError(f"Item {item_name} keywords list is empty.")
+                raise ValueError(f"Item [blue]{item_name}[/blue] keywords list is empty.")
             # if there are other keys in item_config, raise an error
             for key in item_config:
-                if key not in ["keywords", "marketplace", "notify", "exclude_keywords"]:
-                    raise ValueError(f"Item {item_name} contains an invalid key {key}.")
+                if key not in [
+                    "keywords",
+                    "marketplace",
+                    "notify",
+                    "exclude_keywords",
+                    "exclude_sellers",
+                ]:
+                    raise ValueError(
+                        f"Item [blue]{item_name}[/blue] contains an invalid key {key}."
+                    )
 
     def validate_users(self) -> None:
         # check for required fields in each user
@@ -76,7 +93,7 @@ class Config:
                 for user in marketplace_config["notify"]:
                     if user not in self.config["user"]:
                         raise ValueError(
-                            f"User {user} specified in {marketplace_config['name']} does not exist."
+                            f"User [blue]{user}[/blue] specified in [blue]{marketplace_config['name']}[/blue] does not exist."
                         )
 
         # if user is specified for any search item, they must exist
@@ -87,5 +104,5 @@ class Config:
                 for user in item_config["notify"]:
                     if user not in self.config["user"]:
                         raise ValueError(
-                            f"User {user} specified in {item_config['name']} does not exist."
+                            f"User [blue]{user}[/blue] specified in [blue]{item_config['name']}[/blue] does not exist."
                         )