+++ CHANGELOG.md
@@ -1,14 +1,21 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-
 ## [Unreleased]
 
+## [0.2.0] - 2025-01-21
+
+- Allow the definition of a reusable config file from `~/.ai-marketplace-monitor/config.toml`
+- Fix a bug that prevents the sending of phone notification
+
 ## [0.1.0] - 2025-01-20
+
 ### Added
+
 - First release on PyPI.
 
 [Unreleased]: https://github.com/BoPeng/ai-marketplace-monitor/compare/v0.1.0...HEAD